Rename xhcichip header
